requirements.txt
Django==2.0.1
djangorestframework==3.7.7
django-extensions==1.9.9
psycopg2==2.7.3.2